Comisions Function Added

+++ app/models/detalle_gastos.py
@@ -1,5 +1,5 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Float, DateTime
-from datetime import datetime
+from sqlalchemy import Column, Integer, String, ForeignKey, DECIMAL, DateTime
+from datetime import datetime, timezone
 from sqlalchemy.orm import relationship
 from app.models.database import Base
 
@@ -9,6 +9,6 @@ class DetalleGasto(Base):
     id = Column(Integer, primary_key=True, index=True)
     id_trabajo = Column(Integer, ForeignKey("trabajos.id", ondelete="CASCADE"), nullable=False)
     descripcion = Column(String(255), nullable=False)
-    monto = Column(Float, nullable=False)
+    monto = Column(DECIMAL(10, 2), nullable=False)  # Cambiado a DECIMAL para coincidir con la BD
 
     trabajo = relationship("Trabajo", back_populates="detalle_gastos")
